+++ packages/starter/templates/addons/codestyle/.stylelintrc.cjs
@@ -1,4 +1,4 @@
-module.exports = {
+export default {
   extends: ['stylelint-config-standard-scss', 'stylelint-config-prettier-scss', 'stylelint-config-clean-order'],
   plugins: ['stylelint-scss'],
   rules: {