+++ packages/starter/templates/backbone/ssr/src/types/global.declaration.ts
@@ -34,7 +34,7 @@ declare module '*.svg' {
 }
 
 declare module '*.component.svg' {
-  import * as React from 'react';
+  import type * as React from 'react';
 
   const ReactComponent: React.FC<React.SVGProps<SVGSVGElement>>;
   export default ReactComponent;