+++ packages/tester/helpers/create-gql-fetch.js
@@ -15,7 +15,7 @@ const createGqlFetch = (url) => async (query) => {
 
     return await response.json();
   } catch (e) {
-    throw new Error(e);
+    throw new Error(e, { cause: e });
   }
 };