+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [8.0.0]
 
 - Codestyle module has been reworked for better ESLint 10 support
+- Added:
+  - eslint-plugin-no-only-tests
 - TypeScript 6+ fully support
 - Fixed d.ts generation for CSS Modules
 - Claude Code adaptation